Delete mentors, topic, students

+++ graduation_system_app/views/topics.py
@@ -5,6 +5,7 @@ from datetime import datetime
 from django.core.urlresolvers import reverse
 from django.http import HttpRequest
 from django.http import HttpResponse
+from django.http import HttpResponseNotFound
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.template import RequestContext
@@ -56,7 +57,10 @@ def create(request):
                             context_data)
 
 def delete(request, id):
-    topic = Topic.objects.filter(id=id)
-    topic.delete()
+    if request.is_ajax():
+        topic = Topic.objects.filter(id=id)
+        topic.delete()
 
-    return HttpResponse(json.dumps('i did it'), content_type = "application/json")
+        return HttpResponse(json.dumps('Success'), content_type = "application/json")
+
+    return HttpResponseNotFound()
